Added header class and a sticky header
Header class allows for elements within to be in-line with other elements within the header more simply.  Additionally, in order to make the header sticky everything within needs to be enclosed in the same div.  This allows for the header menu to remain at the top of the page even when scrolling